Remove format 14 cmap tool and ucd data, and update Makefile
The tool and data was moved to nototools.

Makefile now needs to invoke the tool as a script, rather than passing
it to python, since the tool is no longer local.  Running setup.py in
nototools will make the script accessible.  Setup (apparently) copies
the script to /usr/local/bin and it is not automatically updated when
nototools changes, still need to figure out how to address this.