Reintroduce check for missing binary, using a different method.
Another way to do this is to create a target, but when I tried that
it reintroduced the dependency analysis problem-- rules with this
dependency always triggered. Of course I might have missed something
subtle.  Both the target approach and this approach come from
http://stackoverflow.com/questions/5618615/check-if-a-program-exists-from-a-makefile)

This always runs, but unless you're just running clean you'd always want to
perform this check, so I think that's ok.