Generate file references relative to destination of output file.
Formerly, we wrote the file paths as given on the command line, the
assumption being that the output file was in the cwd and the paths
to the directories would be correct.

However if we want to generate the output file somewhere other than
the cwd the generated image paths don't work.  This takes the location
of the output file into account and either generates relative paths
if the files are under the output file directory, or absolute paths
otherwise.