Update for new waveflag semantics.
Previously our copy of waveflag took just an input and output filename.
Upstream takes a prefix and one or more input filenames, and concatenates
the prefix to the input filename as the output.

The makefile is changed to pass a prefix and the input filename, instead
of the input filename and the output filename as it formerly did.

Unfortunately for us, our inputs have a directory prefix since they're
not in the current directory, and we don't want this prefix in the output
file path.  So we tweak our copy of waveflag.c to call basename on the
input file path before we append it to the prefix.

We also make the tool a little less noisy by putting more printfs
under the debug flag.