Use full sequences to name images for mixed-gender couples.
We currently name the mixed-gender 'kiss' and 'couple with heart'
images after the single-codepoint sequences.  But aliasing maps
the single codepoint sequence to the gendered sequence, not the
reverse.  As a result the build doesn't create ligatures for the
gendered sequences, since it thinks the image doesn't exist.

Fix this to use the gendered-sequence-names for these images, and
let aliasing work as intended.  This follows the convention we've
adopted of letting the name more completely describe the image
contents, and defining how to represent less-specific sequences
using aliasing rather than baking these decisions into each image
name.