Fix resize of grayscale images.
When imagemagick 6.7.7-10 is processing the '-extent' operator and
discovers an image is grayscale, it turns the 32-bit truecolor sRGB
image into a grayscale image, but does so incorrectly-- the gray
levels and alpha are wrong.  Get around this by using composition
to copy the source image over a slightly larger transparent image.