Identify excluded images explicitly.
Formerly when we were missing an image in the standard emoji sequence
we always labeled it as 'missing'.  However, android excludes some
region flags, so we expect those to be missing and labeling them
implies an error.  Add a new category 'exclude' and populate it
using the unknown-flag alias keys.