Add sign of the horns w/ generic skin tone (1f918)
This is a simple recolor, starting from one of the non-generic sign of
the horns SVGs.

For programs that need both the base and the skin-tone modifier glyph,
this enables both the base (generic) sign of the horns emoji,
and also enables all the emoji versions with specific skin tones.