Move emoji name tool from nototools to here, and change a bit.
The emoji name tool is used to generate titles of emoji images in
pages on the get/noto web site, and group them into related sets.

The original emoji name tool in nototools was based on various
informal emoji name data files plus additional tweaks.  Now there
is a standard source of this data and it is parsed by unicode_data
in nototools.  So we start with that data.

The name data from there is designed for Unicode use and it not
quite appropriate for titles in this format.  This tool tweaked
those names.  This operates in much the same way, except the names
are a bit different when coming from the unicode_data source
and so require somewhat different tweaks.

The emoji alias data/code resides in noto-emoji, but the client for
this tool is in nototools, so it's a bit unclear where to put this.
To avoid circular dependencies, formerly we duplicated some flag
aliasing information in the nototools version.  Now we're trying
the opposite approach, putting the tool in noto-emoji instead...