Add missing SVGs, and clean.
This adds the 50-odd missing SVGs that were not added earlier.
Note that we are still missing the rainbow flag (we don't seem to
have a vector source) and most of the regional flags.  The ~10
regional flags we do have do not match the png images, since they
were sourced differently.  Currently we do not provide the flags
since we cannot rely on support for the svg operations necessary
to implement the 'wave' transform (nor do we have the code to
apply it...)

This also aligns naming between the svg and png.  In particular, it
strips fe0f (emoji variation selector) from the names, and omits
1f48f (kiss) and 1f491 (couple with heart) as these alias.  Those
images are now used for the respective heterosexual couples.

All the files were processed with the revised svg_cleaner tool, and
so are a bit smaller as a result.  Other than some formatting
changes the main changes are removal of the xml doc header, removal
of the application comment, and removal of / changes to some
attributes of the svg tag.