Update SVG images for unicode 11.
These were generated from source, then svg_cleaner was run on
them, and then scour_svg was run on those.

As with the png update there are a lot of small modifications that
don't represent real changes in the images.  This version of scour is
0.37 which might account for some deltas over the previous versions,
or perhaps svg_cleaner accounted for some of these.

Note that two of these images (1f314, 1f318) fail to display properly
on Chrome, Firefox and MacOS finder, though they do display properly
in Adobe Illustrator (go figure).  A later commit will patch these.