Update bad SVG for 'technologist' emoji.
Once again AI export of radial gradients causes problems for
Chrome.  What should have been a white-to-transparent gradient
became a white-to-black 90%-to-100% opaque gradient.  So
hand-edited the SVG to remove the black orb floating in front
of their faces... pity.

Since this was done using guesswork these are not an exact match
for the png versions, but they seem close enough.