Remove unnecessary TTX compile step
Use .ttx with components as input for emoji_builder.py

Originally the .ttx generated by add_glyphs.py was compiled to
a .ttf file so it could be fed to emoji_builder.py. Now we can
feed it a .ttx file and remove the compile step. This way the
original glyph names are preserved and we can start adding
PNGs for the components.