Makefile: make check_sequence a order-only prerequisites
The Makefile rule for the `$(EMOJI).ttf` target passes the first
prerequisite to the emoji_builder python script using the $< automatic
Makefile variable. For this reason, the first prerequisite of this
target must be `$(EMOJI).tmpl.ttf`. Otherwise the following
error message is emitted:

	FileNotFoundError: [Errno 2] No such file or directory: 'check_sequence'

This commit moves the check_sequence prerequisite to the order-only
prerequisites of where it actually belongs as the target itself should
not be updated if check_sequence is executed.