update flag of South Georgia and South Sandwich Islands
The upstream copy on Wikipedia has been updated, so I'm updating it here as well.
The original SVG contained <text> element that referenced an external font (which is not web-safe and that doesn't exist on Wikimedia servers). The text was converted to paths to fix inconsistent renderings on different environments depending on the fallback font used.